Dashboard refactor into smaller components
