+++ src/components/dashboard/tx-list/ScanningTxPlaceholder.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Spinner, View, Text } from '../common'
+import { Spinner, View, Text } from '../../common'
 
 export default class ScanningTxPlaceholder extends React.Component {
   render() {
