+++ src/components/dashboard/tx-list/filter/Filter.ios.js
@@ -3,7 +3,7 @@ import theme from 'metronome-wallet-ui-logic/src/theme'
 import React from 'react'
 import RN from 'react-native'
 
-import { View, BaseBtn } from '../../common'
+import { View, BaseBtn } from '../../../common'
 import DropdownBtn from './DropdownBtn'
 
 class Filter extends React.Component {
