+++ src/components/dashboard/tx-list/row/Amount.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import theme from 'metronome-wallet-ui-logic/src/theme'
 import React from 'react'
 
-import { DisplayValue, Text } from '../../common'
+import { DisplayValue, Text } from '../../../common'
 import ConvertedAmount from './ConvertedAmount'
 import AuctionAmount from './AuctionAmount'
