+++ src/components/dashboard/tx-list/row/AuctionDetails.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Text } from '../../common'
+import { Text } from '../../../common'
 
 export default class AuctionDetails extends React.Component {
   render() {
