+++ src/components/dashboard/tx-list/row/ConvertedDetails.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import { Text } from '../../common'
+import { Text } from '../../../common'
 
 export default class ConvertedDetails extends React.Component {
   static propTypes = {
