+++ src/components/dashboard/tx-list/row/ReceivedDetails.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 
-import { Text } from '../../common'
+import { Text } from '../../../common'
 
 export default class ReceivedDetails extends React.Component {
   static propTypes = {
